requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-limiter==3.3.0
requests==2.26.0
python-dotenv==0.19.0
psutil==5.8.0
google-generativeai==0.3.1 